Regression testing: add test_hybrid
This add generation_test_advanced which search for the appropriate bounds in the
AST.
It starts with each bounds in the range of [0-30].
It then looks for guarded statements and update the bounds in reaction of the
equations encountered in the guard.

Once all the bounds have been update to accommodate to the guards, we generate
the parameters iteration domain which is intersected with the
parameters constraint domain to get the iteration domain of the test.

test_hybrid is a new Make target which launches a textual diff. If the diff test
does not succeed it is followed by a run test on the generated code.

Signed-off-by: Maxime Schmitt <maxime.schmitt91@gmail.com>
[razanajato@etu.unistra.fr: .gitignore, test_hybrid target in top Makefile.am]
Signed-off-by: Harenome Ranaivoarivony Razanajato <razanajato@etu.unistra.fr>