various improvements for test/checker.sh
The previous implementation would decide that a test fails as long as Valgrind
detects memory still in use at exit. The current implementation checks
whether this is due to actual memory leaks and lets the test pass if the
memory is still reachable.

This commit also changes the output of the script `checker.sh`. The output is
more compact as most debug output is now redirected to log files (which
can be found in the directory `test/logs`).

Colour printing now uses `printf` instead of echo: the `-e` option is
useless (and might not even exist) in POSIX-compliant implementations of
`/bin/sh` (such as Dash, for instance) whereas it is necessary with other
shells (such as Bash). With `printf`, both types of shell (at least Dash
and Bash) correctly interpret the escape sequences used for colours.

This also adds support for openscop input in generate_test_advanced.

Signed-off-by: Harenome Ranaivoarivony Razanajato <razanajato@etu.unistra.fr>

test/checker.sh: add support for openscop files

Add support in tests that rely on `generate_test_advanced`.